Tests/LibWeb/Text/expected/textarea-selection-start-selection-end.txt

selectionStart: 0
selectionEnd: 0
selectionStart: 18
selectionEnd: 18
